increase buffer time limit to int16

+++ gframe/single_duel.cpp
@@ -588,7 +588,7 @@ int SingleDuel::Analyze(char* msgbuffer, unsigned int len) {
 		switch (engType) {
 		case MSG_RESET_TIME: {
 			player = BufferIO::ReadInt8(pbuf);
-			int time = BufferIO::ReadInt8(pbuf);
+			short time = BufferIO::ReadInt16(pbuf);
 			if(host_info.time_limit) {
 				if(time)
 					time_limit[player] = time;
