+++ gframe/tag_duel.cpp
@@ -552,7 +552,7 @@ int TagDuel::Analyze(char* msgbuffer, unsigned int len) {
 		switch (engType) {
 		case MSG_RESET_TIME: {
 			player = BufferIO::ReadInt8(pbuf);
-			int time = BufferIO::ReadInt8(pbuf);
+			short time = BufferIO::ReadInt16(pbuf);
 			if(host_info.time_limit) {
 				if(time)
 					time_limit[player] = time;
